requirements.txt
pydantic
openai
typer
rich
loguru
pyyaml
